Move cuDNN tests from test_ltx2_clip_loading.py to test_sdpa.py
